Compare hidden.
