+++ src/MainWindowServices/ButtonService.cs
@@ -1,9 +1,5 @@
-﻿using System.Collections.Generic;
-using System.Collections.ObjectModel;
+﻿using System.Collections.ObjectModel;
 using System.Windows;
-using System.Windows.Controls;
-using CudaHelioCommanderLight.Enums;
-using CudaHelioCommanderLight.Helpers;
 using CudaHelioCommanderLight.Models;
 using CudaHelioCommanderLight.Operations;
 using Microsoft.Win32;
@@ -16,14 +12,6 @@ public class ButtonService
     {
         string message = "Slovak Academy of Sciences\n\nDeveloped by: Martin Nguyen, Pavol Bobik\n\nCopyright 2023";
         MessageBox.Show(message, "About Us", MessageBoxButton.OK, MessageBoxImage.Information);
-    }
-    
-    public void SwitchPanels(PanelType panelType, List<Grid> grids)
-    {
-        foreach (var grid in grids)
-        {
-            grid.Visibility = Visibility.Hidden;
-        }
     }
     
     public void ExportJsonBtn_Click(object sender, RoutedEventArgs e, ObservableCollection<ExecutionDetail> executionDetailList, int executionDetailSelectedIdx)
@@ -52,7 +40,6 @@ public class ButtonService
             ExportAsJsonOperation.Operate(exportModel);
         }
     }
-    
     
     
 }
